pyproject.toml
[project]
name = "research-flow"
version = "0.1.0"
description = "Agentic research system with persistent working memory for ML/DL experiments"
requires-python = ">=3.12"
dependencies = [
    "mlx>=0.31.0",
    "numpy>=2.0",
    "loguru>=0.7",
    "tqdm>=4.60",
    "mlflow>=2.18",
    "pyyaml>=6.0.3",
    "pymupdf>=1.24",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "pre-commit>=4.0",
]
viz = [
    "matplotlib>=3.9",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8",
    "pre-commit>=4.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
